+++ src/mixin/StoreDependencyMixin.js
@@ -3,7 +3,8 @@
  */
 
 var EventHandler = require('../event/EventHandler.js');
-var StoreDependencyDefinition = require('../store/StoreDependencyDefinition.js');
+var StoreDependencyDefinition =
+  require('../store/StoreDependencyDefinition.js');
 var StoreFacade = require('../store/StoreFacade.js');
 
 function havePropsChanged(
@@ -22,7 +23,9 @@ function hasStateChanged(
 ): bool {
   return Object
     .keys(nextState)
-    .some(key => !stores.hasOwnProperty(key) && oldState[key] !== nextState[key]);
+    .some(
+      key => !stores.hasOwnProperty(key) && oldState[key] !== nextState[key]
+    );
 }
 
 function storeChangeCallback(
@@ -50,17 +53,27 @@ function StoreDependencyMixin(
 
   return {
     componentWillMount(): void {
-      var stores = dependencies.getStores();
-      this._storeDependencyHandlers = Object.keys(stores).map(key => {
-        return stores[key].addOnChange(
-          storeChangeCallback.bind(
-            null,
-            this,
-            dependencies,
-            key
-          )
-        );
-      });
+      var i;
+      var key;
+      var store;
+      var storeMap = dependencies.getStores();
+      var stores;
+      this._storeDependencyHandlers = [];
+      for (key in storeMap) {
+        stores = storeMap[key];
+        for (i = 0; i < stores.length; i++) {
+          this._storeDependencyHandlers.push(
+            stores[i].addOnChange(
+              storeChangeCallback.bind(
+                null,
+                this,
+                dependencies,
+                key
+              )
+            )
+          );
+        }
+      }
     },
 
     componentWillUnmount(): void {